api/requirements.txt
fastapi
uvicorn
youtube_dl
pylint
pydantic
moviepy
proglog
pygifsicle
orjson
redis
aiofiles
opencensus-ext-azure
azure-storage-blob
